cleanup code for handling processing routes
